Split Payment resource up into multiple files.
