+++ api/namex/resources/payment/__init__.py
@@ -1,0 +1,5 @@
+from .payment import Payments, Payment, Extra
+from .payment_fees import PaymentFees
+from .payment_invoice import PaymentInvoices, PaymentInvoice
+from .payment_receipt import PaymentReceipt
+from .payment_transaction import PaymentTransactions, PaymentTransaction
