+++ api/namex/resources/payment/api_namespace.py
@@ -1,0 +1,4 @@
+from flask_restplus import Namespace
+
+# Register a local namespace for the NR reserve
+api = Namespace('payments', description='API for Making Payments Using SBC Pay')
